UI Dialogue
Dialogue is stable now but the issue persist.
Lumia-
The first button appears bigger and we need to move the mouse in the button and then out. After that we can select the option.
Flickering does not stop when event ""Quest completed'' activates.
For now this version is the most stable